mock_backend/mock_flask_demo/requirements.txt
amqp==2.3.2
atomicwrites==1.2.0
attrs==18.1.0
bcrypt==3.1.4
billiard==3.5.0.4
blinker==1.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
dotenv-python==0.0.1
elasticsearch==6.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-Elasticsearch==0.2.5
Flask-JWT-Extended==3.12.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
mongoengine==0.15.3
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pycparser==2.18
PyJWT==1.6.4
pymongo==3.7.1
pytest==3.7.3
python-dotenv==0.9.1
pytz==2018.5
rauth==0.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
